backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2

# LangGraph
langgraph==0.2.35

# Azure clients (removed azure-ai-openai - using OpenAI API directly)
azure-search-documents>=11.6.0
azure-ai-formrecognizer>=3.3.0
azure-storage-blob>=12.20.0
azure-identity>=1.17.0

# HTTP / misc
httpx==0.27.0
python-dotenv==1.0.1
pypdf==4.1.0
tiktoken==0.7.0

# Logging & typing
loguru==0.7.2
typing-extensions>=4.12.2

# OpenAI API
openai>=1.0.0
